feat(voice): extract TranscriptionProvider abstraction and add Deepgram backend
- Add TranscriptionProvider interface (electron/services/voice) with a
  hasServerVAD capability flag and shared event types
- Extract the OpenAI Realtime logic into OpenAITranscriptionProvider
  (hasServerVAD: false — keeps the client-side commit cadence)
- Add DeepgramTranscriptionProvider (Nova-3, server-side VAD, raw binary
  frames, KeepAlive, CloseStream drain — no client commit timer)
- Reduce VoiceTranscriptionService to a thin coordinator that selects a
  provider by settings.transcriptionProvider, forwards events, and tears
  down the previous provider on each session
- Add transcriptionProvider + deepgramApiKey to VoiceInputSettings, the
  store schema/defaults, and the preload bridge
- Replace the model-reset migration so it no longer reverts the provider
  on every read; default an unknown provider to "openai" and preserve the
  Deepgram key (DEEPGRAM_API_KEY env override added)
- Add a provider selector + Deepgram key field to the voice settings tab
- Split tests into per-provider suites plus coordinator coverage
